Created a hook to extract create location logic
