+++ src/api/tracker.js
@@ -2,5 +2,5 @@ import axios from 'axios';
 
 
 export default axios.create({
-    baseURL:'http://442704db7698.ngrok.io'
+    baseURL:'http://a0827ad6ce2f.ngrok.io'
 })
